+++ src/index.ts
@@ -25,5 +25,3 @@ export * from './topbar';
 export * from './tree';
 export * from './two-box-multiselect';
 export * from './violation';
-
-export * from './select/select-2';
